Modulised functions, added limiter for number of attempts, fixed non square shape
